refactor: update restartable api (#955)
* feat: update @platfromatic/service restartable api

* debug: ts compiler failing test

* test: increase pluginTimeout for failing test

* test: increase tap timeout for client-cli

* refactor: make loadPlugins a fastify plugin

* test: set keep alive timeout equal 1 for cli test

* test: close app after killing a server in client-cli

* feat: set default keepAliveTimeout to 5s

* deps: update @fastify/restartable up to v2.0.0

+++ packages/service/test/cli/compile.test.mjs
@@ -32,23 +32,28 @@ t.test('should compile typescript plugin', async (t) => {
 
   await cp(testDir, cwd, { recursive: true })
 
-  try {
-    const child = await execa('node', [cliPath, 'compile'], { cwd })
-    t.equal(child.stdout.includes('Typescript compilation completed successfully.'), true)
-  } catch (err) {
-    console.log(err.stdout)
-    console.log(err.stderr)
-    t.fail(err.stderr)
-  }
+  const child = execa('node', [cliPath, 'compile'], { cwd })
 
-  const jsPluginPath = path.join(cwd, 'dist', 'plugin.js')
-  try {
-    await access(jsPluginPath)
-  } catch (err) {
-    t.fail(err)
-  }
+  t.teardown(exitOnTeardown(child))
 
-  t.pass()
+  const splitter = split()
+  child.stdout.pipe(splitter)
+
+  for await (const data of splitter) {
+    const sanitized = stripAnsi(data)
+    if (sanitized.includes('Typescript compilation completed successfully.')) {
+      const jsPluginPath = path.join(cwd, 'dist', 'plugin.js')
+      try {
+        await access(jsPluginPath)
+      } catch (err) {
+        t.fail(err)
+      }
+
+      t.pass()
+      return
+    }
+  }
+  t.fail('should compile typescript plugin with a compile command')
 })
 
 t.test('should compile typescript plugin even if typescript is `false`', async (t) => {
@@ -57,23 +62,28 @@ t.test('should compile typescript plugin even if typescript is `false`', async (
 
   await cp(testDir, cwd, { recursive: true })
 
-  try {
-    const child = await execa('node', [cliPath, 'compile'], { cwd })
-    t.equal(child.stdout.includes('Typescript compilation completed successfully.'), true)
-  } catch (err) {
-    console.log(err.stdout)
-    console.log(err.stderr)
-    t.fail(err.stderr)
-  }
+  const child = execa('node', [cliPath, 'compile'], { cwd })
 
-  const jsPluginPath = path.join(cwd, 'dist', 'plugin.js')
-  try {
-    await access(jsPluginPath)
-  } catch (err) {
-    t.fail(err)
-  }
+  t.teardown(exitOnTeardown(child))
 
-  t.pass()
+  const splitter = split()
+  child.stdout.pipe(splitter)
+
+  for await (const data of splitter) {
+    const sanitized = stripAnsi(data)
+    if (sanitized.includes('Typescript compilation completed successfully.')) {
+      const jsPluginPath = path.join(cwd, 'dist', 'plugin.js')
+      try {
+        await access(jsPluginPath)
+      } catch (err) {
+        t.fail(err)
+      }
+
+      t.pass()
+      return
+    }
+  }
+  t.fail('should compile typescript plugin with a compile command')
 })
 
 t.test('should compile typescript plugin with start command', async (t) => {
@@ -83,8 +93,6 @@ t.test('should compile typescript plugin with start command', async (t) => {
   await cp(testDir, cwd, { recursive: true })
 
   const child = execa('node', [cliPath, 'start'], { cwd })
-
-  t.teardown(exitOnTeardown(child))
 
   const splitter = split()
   child.stdout.pipe(splitter)
@@ -156,7 +164,11 @@ t.test('start command should not compile typescript plugin with errors', async (
     await childProcess
     t.fail('should not compile bad typescript plugin')
   } catch (err) {
-    t.equal(err.stderr.includes('Found 1 error'), true)
+    if (!err.stderr.includes('Found 1 error')) {
+      t.comment(err.stdout)
+      t.comment(err.stderr)
+      t.fail('should throw one ts error')
+    }
     childProcess.kill('SIGINT')
   }
